requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argcomplete==1.10.3
beautifulsoup4==4.8.2
chardet==3.0.4
click==8.1.7
compressed-rtf==1.0.6
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
fastapi==0.112.0
h11==0.14.0
idna==3.7
IMAPClient==2.1.0
lxml==5.3.0
olefile==0.47
pdfminer.six==20191110
pillow==10.4.0
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==0.6.23
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
starlette==0.37.2
textract==1.6.5
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.30.5
xlrd==1.2.0
XlsxWriter==3.2.0
